Move programmes forms to conventions app (#518)
* Move programmes forms to conventions app

* move bailleur form

* rename conventions comments to commentaires

* rename conventions comments to commentaires

* add comments to convention_form formulaires

* rename comments to commentaire in ecoloweb
