+++ conventions/forms/__init__.py
@@ -1,12 +1,19 @@
-from conventions.forms.convention_form_attribution import *
 from conventions.forms.avenant import *
-from conventions.forms.convention_form_commentaires import *
+from conventions.forms.bailleur import *
+from conventions.forms.convention_form_annexes import *
+from conventions.forms.convention_form_attribution import *
+from conventions.forms.convention_form_bailleur import *
+from conventions.forms.convention_form_cadatre import *
 from conventions.forms.convention_form_collectif import *
-from conventions.forms.convention_form_programme import *
-from conventions.forms.convention_number import *
+from conventions.forms.convention_form_commentaires import *
+from conventions.forms.convention_form_edd import *
 from conventions.forms.convention_form_financement import *
-from conventions.forms.convention_form_selection import *
 from conventions.forms.convention_form_foyer_residence_variantes import *
+from conventions.forms.convention_form_logements import *
+from conventions.forms.convention_form_programme import *
+from conventions.forms.convention_form_selection import *
+from conventions.forms.convention_form_type_stationnements import *
+from conventions.forms.convention_number import *
 from conventions.forms.notification import *
 from conventions.forms.resiliation import *
 from conventions.forms.type1and2 import *
