+++ conventions/forms/convention_form_collectif.py
@@ -1,8 +1,16 @@
+"""
+Étape Collectif du formulaire par étape de la convention (type Foyer et Résidence)
+"""
+
 from django import forms
 from django.forms import formset_factory
 
 
 class LotCollectifForm(forms.Form):
+    """
+    Formulaire les informations générales des logements collectifs liées à une convention
+    """
+
     uuid = forms.UUIDField()
     foyer_residence_nb_garage_parking = forms.IntegerField(
         required=False,
@@ -29,6 +37,10 @@ class LotCollectifForm(forms.Form):
 
 
 class LocauxCollectifsForm(forms.Form):
+    """
+    Formulaire définissant un espace collectif d'une convention : une ligne du tableau des logements collectifs
+    """
+
     uuid = forms.UUIDField(required=False)
     type_local = forms.CharField(
         label="",
