+++ conventions/models/convention.py
@@ -53,7 +53,7 @@ class Convention(models.Model):
     )
     # fix me: weird to keep fond_propre here
     fond_propre = models.FloatField(null=True, blank=True)
-    comments = models.TextField(null=True, blank=True)
+    commentaires = models.TextField(null=True, blank=True)
     attached = models.TextField(null=True, blank=True)
     statut = models.CharField(
         max_length=25,
@@ -499,7 +499,7 @@ class Convention(models.Model):
             self,
             exclude=[
                 "avenant_types",
-                "comments",
+                "commentaires",
                 "cree_le",
                 "date_resiliation",
                 "donnees_validees",
