+++ conventions/services/commentaires.py
@@ -3,7 +3,7 @@ from conventions.services import utils
 from conventions.services.conventions import ConventionService
 
 
-class ConventionCommentsService(ConventionService):
+class ConventionCommentairesService(ConventionService):
     form: ConventionCommentForm
     return_status: utils.ReturnStatus = utils.ReturnStatus.ERROR
 
@@ -11,9 +11,9 @@ class ConventionCommentsService(ConventionService):
         self.form = ConventionCommentForm(
             initial={
                 "uuid": self.convention.uuid,
-                "comments": self.convention.comments,
+                "commentaires": self.convention.commentaires,
                 **utils.get_text_and_files_from_field(
-                    "comments", self.convention.comments
+                    "commentaires", self.convention.commentaires
                 ),
                 **utils.get_text_and_files_from_field(
                     "attached", self.convention.attached
@@ -28,9 +28,9 @@ class ConventionCommentsService(ConventionService):
             self.convention.attached = utils.set_files_and_text_field(
                 self.form.cleaned_data["attached_files"],
             )
-            self.convention.comments = utils.set_files_and_text_field(
-                self.form.cleaned_data["comments_files"],
-                self.form.cleaned_data["comments"],
+            self.convention.commentaires = utils.set_files_and_text_field(
+                self.form.cleaned_data["commentaires_files"],
+                self.form.cleaned_data["commentaires"],
             )
             self.convention.save()
             self.return_status = utils.ReturnStatus.SUCCESS
