+++ conventions/tests/services/test_commentaires_service.py
@@ -15,7 +15,7 @@ from core.tests import utils_fixtures
 from users.models import User
 
 
-class ConventionCommentsServiceTests(TestCase):
+class ConventionCommentairesServiceTests(TestCase):
     @classmethod
     def setUpTestData(cls):
         utils_fixtures.create_all()
@@ -24,44 +24,54 @@ class ConventionCommentsServiceTests(TestCase):
         request = HttpRequest()
         convention = Convention.objects.get(numero="0001")
         request.user = User.objects.get(username="fix")
-        self.convention_comments_service = commentaires.ConventionCommentsService(
-            convention=convention, request=request
+        self.convention_commentaires_service = (
+            commentaires.ConventionCommentairesService(
+                convention=convention, request=request
+            )
         )
 
     def test_get(self):
-        self.convention_comments_service.get()
+        self.convention_commentaires_service.get()
         self.assertEqual(
-            self.convention_comments_service.return_status, utils.ReturnStatus.ERROR
+            self.convention_commentaires_service.return_status, utils.ReturnStatus.ERROR
         )
         self.assertIsInstance(
-            self.convention_comments_service.form, ConventionCommentForm
+            self.convention_commentaires_service.form, ConventionCommentForm
         )
         self.assertEqual(
-            self.convention_comments_service.form.initial["uuid"],
-            self.convention_comments_service.convention.uuid,
+            self.convention_commentaires_service.form.initial["uuid"],
+            self.convention_commentaires_service.convention.uuid,
         )
-        comments = json.loads(self.convention_comments_service.convention.comments)
-        text = comments["text"] if "text" in comments else None
-        self.assertEqual(
-            self.convention_comments_service.form.initial["comments"], text
+        commentaires = json.loads(
+            self.convention_commentaires_service.convention.commentaires
         )
-        files = comments["files"] if "files" in comments else None
+        text = commentaires["text"] if "text" in commentaires else None
         self.assertEqual(
-            json.loads(self.convention_comments_service.form.initial["comments_files"]),
+            self.convention_commentaires_service.form.initial["commentaires"], text
+        )
+        files = commentaires["files"] if "files" in commentaires else None
+        self.assertEqual(
+            json.loads(
+                self.convention_commentaires_service.form.initial["commentaires_files"]
+            ),
             files,
         )
 
     def test_save(self):
-        self.convention_comments_service.request.POST["comments"] = ("E" * 5001,)
-        self.convention_comments_service.save()
-        self.assertEqual(
-            self.convention_comments_service.return_status, utils.ReturnStatus.ERROR
+        self.convention_commentaires_service.request.POST["commentaires"] = (
+            "E" * 5001,
         )
-        self.assertTrue(self.convention_comments_service.form.has_error("comments"))
+        self.convention_commentaires_service.save()
+        self.assertEqual(
+            self.convention_commentaires_service.return_status, utils.ReturnStatus.ERROR
+        )
+        self.assertTrue(
+            self.convention_commentaires_service.form.has_error("commentaires")
+        )
 
-        self.convention_comments_service.request.POST = {
-            "comments": "this is a new comment",
-            "comments_files": (
+        self.convention_commentaires_service.request.POST = {
+            "commentaires": "this is a new comment",
+            "commentaires_files": (
                 '{"bbfc7e3a-e0e7-4899-a1e1-fc632c3ea6b0": {"uuid": "bbfc7e3a-e0e7'
                 + '-4899-a1e1-fc632c3ea6b0", "thumbnail": "data:image/png;base64,'
                 + 'BLAH...BLAH...==", "size": 31185, "filename": "acquereur1.png"'
@@ -72,15 +82,18 @@ class ConventionCommentsServiceTests(TestCase):
             ),
         }
 
-        self.convention_comments_service.save()
-        self.convention_comments_service.convention.refresh_from_db()
-        comments = json.loads(self.convention_comments_service.convention.comments)
-        self.assertEqual(
-            self.convention_comments_service.return_status, utils.ReturnStatus.SUCCESS
+        self.convention_commentaires_service.save()
+        self.convention_commentaires_service.convention.refresh_from_db()
+        commentaires = json.loads(
+            self.convention_commentaires_service.convention.commentaires
         )
-        self.assertEqual(comments["text"], "this is a new comment")
         self.assertEqual(
-            comments["files"],
+            self.convention_commentaires_service.return_status,
+            utils.ReturnStatus.SUCCESS,
+        )
+        self.assertEqual(commentaires["text"], "this is a new comment")
+        self.assertEqual(
+            commentaires["files"],
             {
                 "bbfc7e3a-e0e7-4899-a1e1-fc632c3ea6b0": {
                     "uuid": "bbfc7e3a-e0e7-4899-a1e1-fc632c3ea6b0",
