+++ conventions/views/__init__.py
@@ -6,7 +6,7 @@ from conventions.views.convention_form_residence_attribution import *
 from conventions.views.convention_form_bailleur import *
 from conventions.views.convention_form_cadastre import *
 from conventions.views.convention_form_collectif import *
-from conventions.views.convention_form_comments import *
+from conventions.views.convention_form_commentaires import *
 from conventions.views.convention_form_edd import *
 from conventions.views.convention_form_financement import *
 from conventions.views.convention_form_foyer_residence_logements import *
