+++ dataset_staging.json
@@ -299,7 +299,7 @@
       "date_fin_conventionnement": null,
       "financement": "PLUS",
       "fond_propre": null,
-      "comments": null,
+      "commentaires": null,
       "statut": "1. Projet",
       "soumis_le": null,
       "premiere_soumission_le": null,
@@ -338,7 +338,7 @@
       "date_fin_conventionnement": null,
       "financement": "PLS",
       "fond_propre": null,
-      "comments": null,
+      "commentaires": null,
       "statut": "6. Résiliée",
       "soumis_le": null,
       "premiere_soumission_le": null,
@@ -377,7 +377,7 @@
       "date_fin_conventionnement": "2048-06-30",
       "financement": "PLUS",
       "fond_propre": null,
-      "comments": "{\"files\": {}, \"text\": \"\"}",
+      "commentaires": "{\"files\": {}, \"text\": \"\"}",
       "statut": "5. Signée",
       "soumis_le": null,
       "premiere_soumission_le": null,
@@ -416,7 +416,7 @@
       "date_fin_conventionnement": "2045-06-30",
       "financement": "PLAI",
       "fond_propre": null,
-      "comments": "{\"files\": {}, \"text\": \"\"}",
+      "commentaires": "{\"files\": {}, \"text\": \"\"}",
       "statut": "5. Signée",
       "soumis_le": "2022-11-17T16:27:53.812Z",
       "premiere_soumission_le": "2022-11-17T16:27:53.812Z",
@@ -455,7 +455,7 @@
       "date_fin_conventionnement": "2062-06-30",
       "financement": "PLUS",
       "fond_propre": 250000.0,
-      "comments": "{\"files\": {}, \"text\": \"\"}",
+      "commentaires": "{\"files\": {}, \"text\": \"\"}",
       "statut": "3. Corrections requises",
       "soumis_le": "2022-11-09T15:48:53.397Z",
       "premiere_soumission_le": "2022-11-09T15:48:53.397Z",
@@ -494,7 +494,7 @@
       "date_fin_conventionnement": "2055-06-30",
       "financement": "PLAI",
       "fond_propre": 250000.0,
-      "comments": "{\"files\": {}, \"text\": \"\"}",
+      "commentaires": "{\"files\": {}, \"text\": \"\"}",
       "statut": "2. Instruction requise",
       "soumis_le": "2022-11-09T13:01:15.531Z",
       "premiere_soumission_le": "2022-11-09T13:01:15.531Z",
@@ -533,7 +533,7 @@
       "date_fin_conventionnement": "2045-06-30",
       "financement": "PLS",
       "fond_propre": 2450000.0,
-      "comments": "{\"files\": {}, \"text\": \"\"}",
+      "commentaires": "{\"files\": {}, \"text\": \"\"}",
       "statut": "4. A signer",
       "soumis_le": "2022-11-17T15:05:20.777Z",
       "premiere_soumission_le": "2022-11-17T15:05:20.777Z",
@@ -572,7 +572,7 @@
       "date_fin_conventionnement": "2042-06-30",
       "financement": "PLAI",
       "fond_propre": 2300000.0,
-      "comments": "{\"files\": {}, \"text\": \"\"}",
+      "commentaires": "{\"files\": {}, \"text\": \"\"}",
       "statut": "5. Signée",
       "soumis_le": "2022-11-17T15:16:23.595Z",
       "premiere_soumission_le": "2022-11-17T15:16:23.595Z",
@@ -611,7 +611,7 @@
       "date_fin_conventionnement": "2045-06-30",
       "financement": "PLAI",
       "fond_propre": null,
-      "comments": null,
+      "commentaires": null,
       "statut": "1. Projet",
       "soumis_le": "2022-11-17T16:27:53.812Z",
       "premiere_soumission_le": "2022-11-17T16:27:53.812Z",
@@ -652,7 +652,7 @@
       "date_fin_conventionnement": "2047-06-30",
       "financement": "PLUS",
       "fond_propre": null,
-      "comments": null,
+      "commentaires": null,
       "statut": "4. A signer",
       "soumis_le": "2022-11-17T17:41:58.566Z",
       "premiere_soumission_le": "2022-11-17T17:41:58.566Z",
@@ -694,7 +694,7 @@
       "date_fin_conventionnement": "2049-06-30",
       "financement": "PLUS",
       "fond_propre": null,
-      "comments": null,
+      "commentaires": null,
       "statut": "1. Projet",
       "soumis_le": "2022-11-17T17:41:58.566Z",
       "premiere_soumission_le": "2022-11-17T17:41:58.566Z",
