+++ stats/views.py
@@ -11,17 +11,18 @@ from django.db.models import (
 )
 from django.db.models.functions import Substr, Concat
 from apilos_settings.models import Departement
-from bailleurs.forms import BailleurForm
+from conventions.forms import BailleurForm
 
 from bailleurs.models import Bailleur
 from comments.models import Comment
-from conventions.forms import ConventionFinancementForm, ProgrammeForm
-from conventions.models import Convention
-from programmes.forms import (
+from conventions.forms import (
+    ConventionFinancementForm,
     LotLgtsOptionForm,
     ProgrammeCadastralForm,
+    ProgrammeForm,
     ProgrammeEDDForm,
 )
+from conventions.models import Convention
 from programmes.models import Lot, Programme
 from stats.stats_from_sql_raw import average_instruction_delay
 from users.models import User
