+++ templates/conventions/commentaires.html
@@ -22,7 +22,7 @@
         {% include "common/step_progressbar.html"%}
         <div class="fr-container">
             <form method="post" action="">
-                {% include "conventions/common/form_comments.html" %}
+                {% include "conventions/common/form_commentaires.html" %}
                 {% include "conventions/common/form_footer_button.html" %}
             </form>
         </div>
