+++ templates/conventions/recapitulatif/commentaires.html
@@ -1,6 +1,6 @@
 {% load custom_filters %}
 
-{% if convention_form_steps|is_a_step:'ConventionCommentsView' or convention_form_steps|is_a_step:'AvenantCommentsView' %}
+{% if convention_form_steps|is_a_step:'ConventionCommentairesView' or convention_form_steps|is_a_step:'AvenantCommentsView' %}
     <div class="fr-grid-row fr-grid-row--gutters">
         <div class="fr-col-12 fr-col-lg-12">
             <div class="fr-card fr-card--no-arrow">
@@ -13,7 +13,7 @@
                         </div>
                         {% include "conventions/actions/goto.html" %}
                     </div>
-                    {% include "common/display_text_and_files_if_exists.html" with text=convention.comments|get_text_from_textfiles file_list=convention.comments|get_files_from_textfiles %}
+                    {% include "common/display_text_and_files_if_exists.html" with text=convention.commentaires|get_text_from_textfiles file_list=convention.commentaires|get_files_from_textfiles %}
                     {% include "common/display_text_and_files_if_exists.html" with text=convention.attached|get_text_from_textfiles label="Fichiers à joindre à la convention" file_list=convention.attached|get_files_from_textfiles object_field="convention__attached__"|add:convention_uuid%}
                 </div>
             </div>
